requirements.txt
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-bigtable==2.23.1
google-cloud-core==2.4.1
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.64.0
grpcio-status==1.62.2
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pymongo==4.7.2
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
urllib3==2.2.1
Werkzeug==3.0.3
